+++ src/pages/Home.jsx
@@ -1,11 +1,14 @@
+import Nav from '../components/Nav';
 import Hero from '../components/Hero';
 import SectionOne from '../components/homeSectionOne';
 import SectionTwo from '../components/homeSectionTwo';
 import SectionThree from '../components/homeSectionThree';
 
 const Home = () => {
+    
     return (
         <div>
+            <Nav />
             <Hero />
             <SectionOne />
             <SectionTwo />
